# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_CameraNearby_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CameraNearby"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Total rows matching the filter, ignoring pagination.",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[CameraNearby]",
        "type": "object"
      },
      "Page_CameraRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CameraRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Total rows matching the filter, ignoring pagination.",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[CameraRead]",
        "type": "object"
      },
...
      "StreamEndpointRead": {
        "description": "How Models 2-4 reach this camera. Credentials are omitted unless the caller\nholds the streams:credentials scope.",
        "properties": {
          "codec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "h264"
            ],
            "title": "Codec"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_primary": {
            "default": true,
            "title": "Is Primary",
            "type": "boolean"
          },
          "protocol": {
            "title": "Protocol",
            "type": "string"
          },
          "reachability": {
            "description": "public_cdn works on any network; direct_ip needs gateway ports open.",
            "title": "Reachability",
            "type": "string"
          },
          "requires_auth": {
            "default": false,
            "title": "Requires Auth",
            "type": "boolean"
          },
          "resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "1920x1080"
            ],
            "title": "Resolution"
          },
          "url": {
            "examples": [
              "rtsp://103.250.160.189:8554/stream/cam04"
            ],
            "title": "Url",
            "type": "string"
          },
          "verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "required": [
          "id",
          "protocol",
          "url",
          "reachability"
        ],
        "title": "StreamEndpointRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Model 1 \u2014 Centralised CCTV Registry & GIS Foundation.",
    "title": "Sentinel CCTV Registry",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cameras": {
      "get": {
        "operationId": "list_cameras_api_v1_cameras_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Free text over uid, name, address.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free text over uid, name, address.",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "department_ids",
            "required": false,
            "schema": {
              "items": {
                "format": "uuid",
                "type": "string"
              },
              "title": "Department Ids",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "camera_types",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Camera Types",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "statuses",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Statuses",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "ownership_classes",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Ownership Classes",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "district_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CameraRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List and filter cameras",
        "tags": [
          "cameras"
        ]
      },
      "post": {
        "description": "Manual onboarding. Routed through the same ingestion pipeline as CSV upload and vendor sync, so a camera typed into a form gets identical validation, vocabulary resolution and dedupe. Re-submitting the same (department, external_camera_id) updates rather than duplicating.",
        "operationId": "create_camera_api_v1_cameras_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CameraCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CameraRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register a single camera",
        "tags": [
          "cameras"
        ]
      }
    },
    "/api/v1/cameras/export.csv": {
      "get": {
        "description": "The same filter as the list endpoint and the tiles, rendered as a spreadsheet a department can reconcile against its own records. An empty result still returns a header row.",
        "operationId": "export_csv_api_v1_cameras_export_csv_get",
        "parameters": [
          {
            "description": "Free text over uid, name, address.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free text over uid, name, address.",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "department_ids",
            "required": false,
            "schema": {
              "items": {
                "format": "uuid",
                "type": "string"
              },
              "title": "Department Ids",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "camera_types",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Camera Types",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "statuses",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Statuses",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "ownership_classes",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Ownership Classes",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "district_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/csv": {}
            },
            "description": "Matching cameras."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export the filtered result set as CSV",
        "tags": [
          "cameras"
        ]
      }
    },
    "/api/v1/cameras/nearby": {
      "get": {
        "description": "The incident-response query: given an FIR location, which cameras could have seen it, closest first. Distances are true metres on the spheroid, not degrees.",
        "operationId": "cameras_nearby_api_v1_cameras_nearby_get",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "lon",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Lon",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "radius_m",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "maximum": 200000,
              "title": "Radius M",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CameraNearby_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cameras within a radius, nearest first",
        "tags": [
          "cameras"
        ]
      }
    },
    "/api/v1/cameras/{camera_id}": {
      "get": {
        "operationId": "get_camera_api_v1_cameras__camera_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "camera_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Camera Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CameraRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Camera",
        "tags": [
          "cameras"
        ]
      }
    },
    "/api/v1/cameras/{camera_id}/audit": {
      "get": {
        "description": "Every recorded change, newest first, with the state before and after.",
        "operationId": "camera_audit_api_v1_cameras__camera_id__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "camera_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Camera Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Camera Audit Api V1 Cameras  Camera Id  Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change history for one camera",
        "tags": [
          "cameras"
        ]
      }
    },
    "/api/v1/cameras/{camera_id}/streams": {
      "get": {
        "description": "Entry point for Models 2-4. Prefer the endpoint whose reachability matches your network: public_cdn works anywhere, direct_ip needs gateway ports open.",
        "operationId": "get_camera_streams_api_v1_cameras__camera_id__streams_get",
        "parameters": [
          {
            "in": "path",
            "name": "camera_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Camera Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StreamEndpointRead"
                  },
                  "title": "Response Get Camera Streams Api V1 Cameras  Camera Id  Streams Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream endpoints for this camera",
        "tags": [
          "cameras"
        ]
      }
    },
...
  }
}